mobile analysis and docs update

+++ periculum/src/main/java/com/periculum/internal/models/IncomeAnalysis.kt
@@ -3,14 +3,14 @@ package com.periculum.internal.models
 data class IncomeAnalysis (
     val averagePredictedSalary: Double? = null,
     val isSalaryEarner: String? = null,
-    val expectedSalaryPaymentDay: Long? = null,
+    val expectedSalaryPaymentDay: Double? = null,
     val frequencyOfSalaryPayments: String? = null,
     val lastDateOfSalaryPayment: String? = null,
-    val numberOfSalaryPayments: Long? = null,
+    val numberOfSalaryPayments: Double? = null,
     val hasOtherIncome: String? = null,
-    val averageOtherIncome: Long? = null,
-    val numberOfOtherIncomePayments: Long? = null,
-    val netAverageMonthlyEarning: Long? = null,
-    val lowestSalary: Long? = null,
+    val averageOtherIncome: Double? = null,
+    val numberOfOtherIncomePayments: Double? = null,
+    val netAverageMonthlyEarning: Double? = null,
+    val lowestSalary: Double? = null,
     val mostRecentSalary: String? = null
 )
